refactor(meeting_backend): extract close + send_message impl methods (#1266)
Continue meeting_backend/mod.rs trim (#1364):

- closing.rs (184 LOC): impl block for close()
- messaging.rs (91 LOC): impl block for send_message()

mod.rs: 587 -> 344 ✓ under the 400-LOC ceiling

All 81 meeting_backend tests pass.
