+++ src/meeting_backend/mod.rs
@@ -13,6 +13,9 @@ mod tests_persist;
 mod tests_persist_extra;
 pub mod types;
 
+mod closing;
+mod messaging;
+
 use std::fmt::Write as _;
 use std::sync::mpsc;
 use std::time::Duration;
@@ -31,7 +34,7 @@ pub use types::{
 };
 
 /// Maximum messages kept in conversation history.
-const MAX_HISTORY: usize = 500;
+pub(super) const MAX_HISTORY: usize = 500;
 
 /// Number of recent messages included verbatim in the LLM prompt.
 const RECENT_WINDOW: usize = 30;
@@ -90,251 +93,6 @@ impl MeetingBackend {
         }
     }
 
-    /// Send a user message and get Simard's response.
-    ///
-    /// Appends both the user message and the assistant response to history.
-    /// The full conversation context is sent to the LLM on each turn.
-    #[tracing::instrument(skip(self), fields(input_len = user_input.len()))]
-    pub fn send_message(&mut self, user_input: &str) -> SimardResult<MeetingResponse> {
-        if !self.is_open {
-            return Err(SimardError::ActionExecutionFailed {
-                action: "send-message".to_string(),
-                reason: "meeting session is closed".to_string(),
-            });
-        }
-
-        let trimmed = user_input.trim();
-        if trimmed.is_empty() {
-            return Ok(MeetingResponse {
-                content: String::new(),
-                message_count: self.history.len(),
-            });
-        }
-
-        // Append user message
-        self.push_message(Role::User, trimmed.to_string());
-
-        // Build the prompt preamble from conversation history
-        let preamble = self.build_conversation_preamble();
-
-        let turn_input = BaseTypeTurnInput {
-            objective: trimmed.to_string(),
-            identity_context: self.system_prompt.clone(),
-            prompt_preamble: preamble,
-        };
-
-        info!(
-            topic = self.topic,
-            messages = self.history.len(),
-            input_len = trimmed.len(),
-            "Sending message to LLM agent…"
-        );
-        let start = std::time::Instant::now();
-
-        let outcome = match self.agent.run_turn(turn_input) {
-            Ok(o) => {
-                info!(
-                    elapsed_ms = start.elapsed().as_millis() as u64,
-                    response_len = o.execution_summary.len(),
-                    "LLM agent returned response"
-                );
-                o
-            }
-            Err(e) => {
-                warn!(elapsed_ms = start.elapsed().as_millis() as u64, error = %e, "LLM agent returned error");
-                return Err(e);
-            }
-        };
-        let response_text = {
-            let extracted = extract_response(&outcome);
-            if extracted.trim().is_empty() {
-                EMPTY_RESPONSE_SENTINEL.to_string()
-            } else {
-                extracted
-            }
-        };
-
-        // Append assistant response
-        self.push_message(Role::Assistant, response_text.clone());
-
-        // Auto-save transcript after every turn so killed meetings don't lose data
-        self.auto_save_transcript();
-
-        debug!(messages = self.history.len(), "Meeting turn completed");
-
-        Ok(MeetingResponse {
-            content: response_text,
-            message_count: self.history.len(),
-        })
-    }
-
-    /// Close the meeting session: summarize, extract action items, link goals,
-    /// auto-export markdown report, persist, and store to memory.
-    ///
-    /// Returns a `MeetingSummary` with the LLM-generated summary text and
-    /// structured action items.
-    #[tracing::instrument(skip(self))]
-    pub fn close(&mut self) -> SimardResult<MeetingSummary> {
-        if !self.is_open {
-            return Err(SimardError::ActionExecutionFailed {
-                action: "close-meeting".to_string(),
-                reason: "meeting session is already closed".to_string(),
-            });
-        }
-
-        self.is_open = false;
-        let duration_secs = self.elapsed_secs();
-
-        // Generate summary via LLM (internal prompt, not visible to operator)
-        let summary_text = self.generate_summary();
-
-        // ── Structured action-item extraction ──
-        let mut action_items = persist::extract_action_items(&self.history);
-
-        // ── Goal linkage ──
-        let goal_titles = self.load_active_goal_titles();
-        if !goal_titles.is_empty() {
-            persist::link_action_items_to_goals(&mut action_items, &goal_titles);
-        }
-
-        // ── Decision extraction ──
-        let decisions = persist::extract_decisions(&self.history);
-
-        // Write JSON transcript to ~/.simard/meetings/
-        let transcript = MeetingTranscript {
-            topic: self.topic.clone(),
-            started_at: self.started_at.clone(),
-            closed_at: Utc::now().to_rfc3339(),
-            duration_secs,
-            summary: summary_text.clone(),
-            messages: self.history.clone(),
-        };
-        let transcript_path = match persist::write_transcript(&transcript) {
-            Ok(p) => Some(p.to_string_lossy().to_string()),
-            Err(e) => {
-                warn!("Failed to write transcript: {e}");
-                None
-            }
-        };
-
-        // Write MeetingHandoff artifact for OODA integration
-        if let Err(e) = persist::write_handoff(
-            &self.topic,
-            &summary_text,
-            &self.history,
-            &action_items,
-            &decisions,
-        ) {
-            warn!("Failed to write meeting handoff: {e}");
-        }
-
-        // ── Extract open questions and themes for the summary ──
-        let open_questions: Vec<String> = persist::extract_open_questions(&self.history)
-            .into_iter()
-            .map(|q| q.text)
-            .collect();
-        // Explicit /theme entries come first; inferred themes fill in the rest.
-        let inferred_themes = persist::extract_themes(&self.history);
-        let mut themes: Vec<String> = self.themes.clone();
-        for t in inferred_themes {
-            let lower = t.to_lowercase();
-            if !themes.iter().any(|e| e.to_lowercase() == lower) {
-                themes.push(t);
-            }
-        }
-
-        // Collect unique participants from messages and action item assignees.
-        let mut participants: Vec<String> = Vec::new();
-        for msg in &self.history {
-            let role_name = match msg.role {
-                Role::User => "operator",
-                Role::Assistant => "simard",
-                Role::System => "system",
-            };
-            let s = role_name.to_string();
-            if !participants.contains(&s) {
-                participants.push(s);
-            }
-        }
-        for a in &action_items {
-            if let Some(ref assignee) = a.assignee
-                && !participants.contains(assignee)
-            {
-                participants.push(assignee.clone());
-            }
-        }
-
-        // ── Auto-export markdown report on /end ──
-        // Convert extracted decision strings to MeetingDecision structs (with rationale).
-        let structured_decisions: Vec<crate::meeting_facilitator::MeetingDecision> = decisions
-            .iter()
-            .map(|d| {
-                let rationale = persist::extract_decision_rationale_pub(d, &self.history);
-                let participants = persist::extract_decision_participants_pub(d, &self.history);
-                crate::meeting_facilitator::MeetingDecision {
-                    description: d.clone(),
-                    rationale,
-                    participants,
-                }
-            })
-            .collect();
-        let markdown_report_path = match persist::write_handoff_markdown_report(
-            &self.topic,
-            &self.started_at,
-            &summary_text,
-            &self.history,
-            &action_items,
-            &structured_decisions,
-        ) {
-            Ok(p) => Some(p.to_string_lossy().to_string()),
-            Err(e) => {
-                warn!("Failed to write handoff markdown report: {e}");
-                None
-            }
-        };
-
-        // ── Memory consolidation ──
-        if let Some(ref bridge) = self.bridge {
-            persist::store_enriched_cognitive_memory(
-                &**bridge,
-                &self.topic,
-                &summary_text,
-                &self.history,
-                &action_items,
-                &decisions,
-            );
-        }
-
-        // Close the agent session
-        if let Err(e) = self.agent.close() {
-            warn!("Failed to close agent session: {e}");
-        }
-
-        info!(
-            topic = self.topic,
-            messages = self.history.len(),
-            action_items = action_items.len(),
-            decisions = decisions.len(),
-            duration_secs,
-            "Meeting session closed with structured handoff"
-        );
-
-        Ok(MeetingSummary {
-            topic: self.topic.clone(),
-            summary_text,
-            message_count: self.history.len(),
-            duration_secs,
-            transcript_path,
-            action_items,
-            decisions,
-            markdown_report_path,
-            open_questions,
-            themes,
-            participants,
-        })
-    }
-
-    /// Get current session status.
     pub fn status(&self) -> SessionStatus {
         SessionStatus {
             topic: self.topic.clone(),
